Import fx2lib into fx2lafw directly.
We want a local fx2lib copy in fx2lafw for various reasons, e.g. to make
life simpler for all distros (none of which ship any fx2lib packages,
and neither would it make sense to do so really), and as we'll want/need
local changes to fix build issues and possibly other stuff anyway.

This copy of fx2lib is not the main upstream version from

  git://github.com/mulicheng/fx2lib.git

which has been last updated in 2010, but rather from

  git://github.com/ubixum/fx2lib.git

The only difference is that the latter contains one patch which fixes
some issues with more recent sdcc versions (interrupt vs. __interrupt,
_asm vs. __asm, xdata vs. __xdata__, and many others).

The imported fx2lib copy is "416c104cbd" from Feb 12, 2012.