Support older Cypress FX2 (not FX2LP) too.
The Cypress FX2 chip is an older version of the FX2LP chip with a few
differences that might be relevant for fx2lafw, the major one being
that is has less internal SRAM.

We adapt the linking step to work for both FX2 and FX2LP, thus
keeping only one set of *.fw files for both chip versions.

Changes to the fx2lafw driver in libsigrok are not required.

Some logic analyzer devices such as older USBee SX versions or more recent
devices which use a "fake" FX2LP (which is actually an FX2) need this.

This closes bug #3.