Add FX2 firmware for the Hantek 6022BE USB scope.
This firmware was originally created by Jochen Hoenicke in 05/2015
for the Hantek6022API github project by Robert Cope. Full credit for
the implementation of the firmware goes to Jochen Hoenicke, thanks!

Github project:

  https://github.com/rpcope1/Hantek6022API

We're incorporating the firmware into the sigrok-firmware-fx2lafw
repo/tarball for convenience of sigrok users. E.g. so that they can
easily get it with the other fx2lafw firmware files from distro
packages, the build scripts in sigrok-util automatically build/install
it, the sigrok Windows installers automatically ship it, the sigrok
Android APKs automatically ship it, etc. etc.

The files we're integrating into fx2lafw from Hantek6022API are taken from
the PyHT6022/HantekFirmware/custom directory of that repo, using the state
as of 03/2016 (the files were last modified in 05/2015, though).

The git hash of the last relevant commit there was:

  0498e2ab239aabb1084c19e1e24faf56764b217e

Files integrated into fx2lafw:

 - hw/hantek-6022be/dscr.a51: Copied unmodified.

 - hantek_6022be.c: This file is created by appending the
   contents of the original device.c and fw.c together without
   other modifications.

The license of both files is LGPL 2.1 (or later).