fx2lafw: Delay start of GPIF until first BULK IN request is received
Apparently the host controller may issue the CONTROL transfer with
the START command earlier than any of the BULK IN requests. The result
is a filled up EP2 packet FIFO, which aborts the acquisition.

Waiting for the first BULK IN request matches the sample code in the
TRM, see e.g. Figure 10-30.

This fixes bug #655.